LGMVIP-Web-Task-Number 1
To-do lists are a list of tasks that an individual needs to complete or accomplish. Tasks are typically put in order by priority or importance. A to-do list can be written on a piece of paper or by utilizing task management software. Items on the list can pertain to one's personal or professional life. I learn to make a To-Do List App or Website with HTML, CSS and JavaScript. In this To Do list we can add our task list and mark that task as completed or incomplete. We can write the new task in the input box and click on the "Add" button then the new task will be added to the list. We can click on the task name to mark it as completed or incomplete task. We will add a close icon to delete the particular task from the ToDo list.

In this JavaScript Project we are using Local Storage that will store the task list or To Do list in your web browser, So that if you close and restart the browser then also it will display your saved To-Do List